move VDU handling into VDUStreamProcessor
Lots of moving around of functions so they’re in more logical places

Audio is particularly re-arranged, so vdu_audio file is now solely about handling VDU stream

Some minor tweaking of audio code to do some slightly better memory handling (unique_ptr instead of shared_ptr for things that should be unique)
